chore(next.clayui.com): remove template components
