+++ next.clayui.com/gatsby/createPages.js
@@ -22,10 +22,6 @@ const slugWithBar = path => {
 const createDocs = (actions, edges, mdx, blacklist = []) => {
 	const {createPage, createRedirect} = actions;
 	const docsTemplate = path.resolve(__dirname, '../src/templates/docs.js');
-	const componentsTemplate = path.resolve(
-		__dirname,
-		'../src/templates/components.js'
-	);
 	const blogTemplate = path.resolve(__dirname, '../src/templates/blog.js');
 
 	edges
@@ -87,8 +83,6 @@ const createDocs = (actions, edges, mdx, blacklist = []) => {
 
 				if (slug.includes('blog/')) {
 					template = blogTemplate;
-				} else if (slug.includes('docs/components')) {
-					template = componentsTemplate;
 				} else if (slug.includes('docs/')) {
 					template = docsTemplate;
 				}
